require not replace
As was done in solo-io/gloo, we should be using exclude directives that
mirror the deprecate functionality in go1.18. This block has been
borrowed wholesale from that repo.

As well, now instead of replacing k8s.io/client-go with the version that
we actually want, we can just require it. Other requires have been left
unmodified.

Updates #461.