add LICENSE
This repo lacks a license file. This means that using this code is a
legally grey imposition, thus other tools have opted to simply not:
https://pkg.go.dev/github.com/solo-io/solo-apis/pkg/api/gloo.solo.io/v1.

This seems to be an oversight during the migration out of the old
monorepo, but I could be wrong. Either way, I have copied the LICENSE
file from there, since this code at least was licensed under Apache-2.0.